app/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
python-docx>=1.1.0
python-multipart>=0.0.6

langchain==0.3.9
langgraph==0.2.56
langchain-openai>=0.0.8
langchain-core~=0.3.60

numpy>=1.24
huggingface_hub>=0.20
sentence-transformers>=2.2
faster-whisper>=1.0
ctranslate2>=4.3
pyannote.audio>=3.1
torch==2.3.0
torchaudio>=2.2

mediapipe>=0.10.21
opencv-python>=4.10
moviepy>=1.0.3

bcrypt>=4.0.1
PyJWT>=2.8.0
email-validator>=2.0.0
google-generativeai>=0.3.2
PyMuPDF>=1.23.0
Pillow>=10.0.0
python-pptx>=0.6.21
pyairtable>=2.3.3
